#1835 Extracted java-dependent code to a separate module with optional java dependency
