+++ java/src/main/java/org/ice1000/julia/java/ui/configuration/JuliaModuleConfigEditorProvider.kt
@@ -1,0 +1,34 @@
+/*
+ *     Julia language support plugin for Intellij-based IDEs.
+ *     Copyright (C) 2024 julia-intellij contributors
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.ice1000.julia.java.ui.configuration
+
+import com.intellij.openapi.module.ModuleConfigurationEditor
+import com.intellij.openapi.roots.ui.configuration.ContentEntriesEditor
+import com.intellij.openapi.roots.ui.configuration.ModuleConfigurationEditorProvider
+import com.intellij.openapi.roots.ui.configuration.ModuleConfigurationState
+
+class JuliaModuleConfigEditorProvider : ModuleConfigurationEditorProvider {
+    override fun createEditors(state: ModuleConfigurationState): Array<ModuleConfigurationEditor> {
+        val module = state.currentRootModel?.module ?: return emptyArray()
+        return arrayOf(
+            ContentEntriesEditor(module.name, state),
+            JuliaCompileOutputEditor(state)
+        )
+    }
+}
